+++ src/collection.rs
@@ -17,7 +17,6 @@ use crate::{
 pub trait MenuItemCollection<R> {
     fn bounds_of(&self, nth: usize) -> Rectangle;
     fn title_of(&self, nth: usize) -> &str;
-    fn details_of(&self, nth: usize) -> &str;
     fn value_of(&self, nth: usize) -> R;
     fn interact_with(&mut self, nth: usize) -> R;
     /// Whether an item is selectable. If not, the item will be skipped.
@@ -64,11 +63,6 @@ where
     fn title_of(&self, nth: usize) -> &str {
         debug_assert!(nth == 0);
         self.title()
-    }
-
-    fn details_of(&self, nth: usize) -> &str {
-        debug_assert!(nth == 0);
-        self.details()
     }
 
     fn count(&self) -> usize {
@@ -142,10 +136,6 @@ where
 
     fn title_of(&self, nth: usize) -> &str {
         self.items.as_ref()[nth].title()
-    }
-
-    fn details_of(&self, nth: usize) -> &str {
-        self.items.as_ref()[nth].details()
     }
 
     fn selectable(&self, nth: usize) -> bool {
@@ -241,10 +231,6 @@ where
         self.object.title_of(nth)
     }
 
-    fn details_of(&self, nth: usize) -> &str {
-        self.object.details_of(nth)
-    }
-
     fn selectable(&self, nth: usize) -> bool {
         self.object.selectable(nth)
     }
@@ -308,15 +294,6 @@ where
             self.parent.title_of(nth)
         } else {
             self.object.title_of(nth - count)
-        }
-    }
-
-    fn details_of(&self, nth: usize) -> &str {
-        let count = self.parent.count();
-        if nth < count {
-            self.parent.details_of(nth)
-        } else {
-            self.object.details_of(nth - count)
         }
     }